+++ src/rtpsession_priv.h
@@ -66,7 +66,7 @@ void update_avg_rtcp_size(RtpSession *session, int bytes);
 mblk_t * rtp_session_network_simulate(RtpSession *session, mblk_t *input, bool_t *is_rtp_packet);
 void ortp_network_simulator_destroy(OrtpNetworkSimulatorCtx *sim);
 
-void rtcp_common_header_init(rtcp_common_header_t *ch, RtpSession *s,int type, int rc, int bytes_len);
+void rtcp_common_header_init(rtcp_common_header_t *ch, RtpSession *s,int type, int rc, size_t bytes_len);
 
 mblk_t * make_xr_rcvr_rtt(RtpSession *session);
 mblk_t * make_xr_dlrr(RtpSession *session);
